chore: updated to new ftb library layout

+++ common/src/main/java/dev/ftb/mods/ftbquests/quest/ServerQuestFile.java
@@ -1,16 +1,29 @@
 package dev.ftb.mods.ftbquests.quest;
 
+import net.minecraft.core.HolderLookup;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.storage.LevelResource;
 import com.mojang.util.UndashedUuid;
+
 import dev.architectury.networking.NetworkManager;
 import dev.architectury.platform.Platform;
 import dev.architectury.utils.Env;
+
 import dev.ftb.mods.ftblibrary.snbt.SNBT;
-import dev.ftb.mods.ftblibrary.snbt.SNBTCompoundTag;
 import dev.ftb.mods.ftblibrary.util.NetworkHelper;
 import dev.ftb.mods.ftbquests.FTBQuests;
 import dev.ftb.mods.ftbquests.events.QuestProgressEventData;
 import dev.ftb.mods.ftbquests.integration.PermissionsHelper;
-import dev.ftb.mods.ftbquests.net.*;
+import dev.ftb.mods.ftbquests.net.CreateOtherTeamDataMessage;
+import dev.ftb.mods.ftbquests.net.DeleteObjectResponseMessage;
+import dev.ftb.mods.ftbquests.net.MoveChapterGroupResponseMessage;
+import dev.ftb.mods.ftbquests.net.SyncEditorPermissionMessage;
+import dev.ftb.mods.ftbquests.net.SyncQuestsMessage;
+import dev.ftb.mods.ftbquests.net.SyncTeamDataMessage;
+import dev.ftb.mods.ftbquests.net.TeamDataChangedMessage;
+import dev.ftb.mods.ftbquests.net.TeamDataUpdate;
 import dev.ftb.mods.ftbquests.quest.reward.RewardType;
 import dev.ftb.mods.ftbquests.quest.reward.RewardTypes;
 import dev.ftb.mods.ftbquests.quest.task.Task;
@@ -25,11 +38,6 @@ import dev.ftb.mods.ftbteams.api.event.PlayerChangedTeamEvent;
 import dev.ftb.mods.ftbteams.api.event.PlayerLoggedInAfterTeamEvent;
 import dev.ftb.mods.ftbteams.api.event.TeamCreatedEvent;
 import dev.ftb.mods.ftbteams.data.PartyTeam;
-import net.minecraft.core.HolderLookup;
-import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.storage.LevelResource;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -38,16 +46,20 @@ import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Stream;
+import org.jspecify.annotations.Nullable;
 
 public class ServerQuestFile extends BaseQuestFile {
 	public static final LevelResource FTBQUESTS_DATA = new LevelResource("ftbquests");
 
+	@Nullable
 	public static ServerQuestFile INSTANCE;
 
 	public final MinecraftServer server;
 	private boolean shouldSave;
 	private boolean isLoading;
+	@Nullable
 	private Path folder;
+	@Nullable
 	private ServerPlayer currentPlayer = null;
 
 	public ServerQuestFile(MinecraftServer s) {
@@ -130,6 +142,7 @@ public class ServerQuestFile extends BaseQuestFile {
 	}
 
 	@Override
+	@Nullable
 	public Path getFolder() {
 		return folder;
 	}
@@ -144,7 +157,9 @@ public class ServerQuestFile extends BaseQuestFile {
 			object.deleteSelf();
 			refreshIDMap();
 			markDirty();
-			object.getPath().ifPresent(path -> FileUtils.delete(getFolder().resolve(path).toFile()));
+			if (getFolder() != null) {
+				object.getPath().ifPresent(path -> FileUtils.delete(getFolder().resolve(path).toFile()));
+			}
 		}
 
 		NetworkHelper.sendToAll(server, new DeleteObjectResponseMessage(id));
@@ -174,11 +189,12 @@ public class ServerQuestFile extends BaseQuestFile {
 		deleteSelf();
 	}
 
+	@Nullable
 	public ServerPlayer getCurrentPlayer() {
 		return currentPlayer;
 	}
 
-	public void withPlayerContext(ServerPlayer player, Runnable toDo) {
+	public void withPlayerContext(@Nullable ServerPlayer player, Runnable toDo) {
 		currentPlayer = player;
 		try {
 			toDo.run();
